+++ consts.go
@@ -2,4 +2,4 @@ package binance_connector
 
 const Name = "binance-connector-go"
 
-const Version = "0.2.0"
+const Version = "0.3.0"